Extract updateOwnerReferences to util package
This functionality will be re-used by ExtendedJob

[#163483576](https://www.pivotaltracker.com/story/show/163483576)
