+++ pkg/kube/apis/apis.go
@@ -1,5 +1,10 @@
 package apis
 
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+)
+
 const (
 	// GroupName defines the kube group name for all controllers
 	// It's what's used when you specify the apiVersion of a resource.
@@ -11,3 +16,11 @@ const (
 	//   ...
 	GroupName = "fissile.cloudfoundry.org"
 )
+
+// Object is used as a helper interface when passing Kubernetes resources
+// between methods.
+// All Kubernetes resources should implement both of these interfaces
+type Object interface {
+	runtime.Object
+	metav1.Object
+}
