+++ pkg/kube/apis/extendedstatefulset/v1alpha1/types.go
@@ -9,7 +9,6 @@ import (
 	"github.com/pkg/errors"
 	"k8s.io/api/apps/v1beta2"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 
 	"code.cloudfoundry.org/cf-operator/pkg/kube/apis"
 )
@@ -26,14 +25,6 @@ var (
 	// AnnotationVersion is the annotation key for the StatefulSet version
 	AnnotationVersion = fmt.Sprintf("%s/version", apis.GroupName)
 )
-
-// Object is used as a helper interface when passing Kubernetes resources
-// between methods.
-// All Kubernetes resources should implement both of these interfaces
-type Object interface {
-	runtime.Object
-	metav1.Object
-}
 
 // ExtendedStatefulSetSpec defines the desired state of ExtendedStatefulSet
 type ExtendedStatefulSetSpec struct {
