+++ pkg/kube/controllers/extendedstatefulset/reconciler_test.go
@@ -20,6 +20,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
+	"code.cloudfoundry.org/cf-operator/pkg/kube/apis"
 	exss "code.cloudfoundry.org/cf-operator/pkg/kube/apis/extendedstatefulset/v1alpha1"
 	"code.cloudfoundry.org/cf-operator/pkg/kube/controllers"
 	exssc "code.cloudfoundry.org/cf-operator/pkg/kube/controllers/extendedstatefulset"
@@ -398,7 +399,7 @@ var _ = Describe("ReconcileExtendedStatefulSet", func() {
 							BlockOwnerDeletion: helper.Bool(true),
 						}
 
-						for _, obj := range []exss.Object{configMap1, configMap2, secret1, secret2} {
+						for _, obj := range []apis.Object{configMap1, configMap2, secret1, secret2} {
 							Expect(obj.GetOwnerReferences()).ShouldNot(ContainElement(ownerRef))
 						}
 					})
@@ -473,7 +474,7 @@ var _ = Describe("ReconcileExtendedStatefulSet", func() {
 								BlockOwnerDeletion: helper.Bool(true),
 							}
 
-							for _, obj := range []exss.Object{configMap1, secret1} {
+							for _, obj := range []apis.Object{configMap1, secret1} {
 								Expect(obj.GetOwnerReferences()).Should(ContainElement(ownerRef))
 							}
 						})
@@ -549,7 +550,7 @@ var _ = Describe("ReconcileExtendedStatefulSet", func() {
 								BlockOwnerDeletion: helper.Bool(true),
 							}
 
-							for _, obj := range []exss.Object{configMap1, configMap2, secret1, secret2} {
+							for _, obj := range []apis.Object{configMap1, configMap2, secret1, secret2} {
 								Expect(obj.GetOwnerReferences()).Should(ContainElement(ownerRef))
 							}
 						})
